+++ src/community/components/form/text-editor/text-editor.stories.tsx
@@ -1,4 +1,4 @@
-import { Meta, StoryFn, StoryObj } from '@storybook/react';
+import { Meta, StoryFn, StoryObj } from '@storybook/react-vite';
 import dompurify from 'dompurify';
 import React from 'react';
 import * as showdown from 'showdown';