+++ src/community/components/table-of-contents/table-of-contents.stories.tsx
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker';
-import { Meta, StoryFn, StoryObj } from '@storybook/react';
+import { Meta, StoryFn, StoryObj } from '@storybook/react-vite';
 import React from 'react';
 
 import { Heading } from '../../../tedi/components/base/typography/heading/heading';
@@ -26,8 +26,8 @@ const meta: Meta<TableOfContentsProps> = {
       },
     },
     layout: 'fullscreen',
-    backgrounds: { default: 'subtle' },
   },
+  globals: { backgrounds: { value: 'subtle' } },
 };
 
 export default meta;