+++ src/tedi/components/content/carousel/index.ts
@@ -1,0 +1,7 @@
+export * from './carousel';
+export * from './components/carousel-content/carousel-content';
+export * from './components/carousel-footer/carousel-footer';
+export * from './components/carousel-header/carousel-header';
+export * from './components/carousel-indicators/carousel-indicators';
+export * from './components/carousel-navigation/carousel-navigation';
+export * from './carousel-utils';